Access the application view model
This adds the ability to globally access the application View Model. You
can access it with:

```js
var state = $("html").viewModel();
```

This is useful when developing, you might want to inspect the
Application View Model from your development tools to see what values it
has, or to change values to see how the application changes as a
response.