Support H2 headers object
Node's `http2` module has a slightly different API than `http` and
`https`. Instead of a `request` object, it provides you a POJO
containing headers.

In SSR we use the request object to check that there is a matching route
rule (to support 404s). This makes it possible for autorender to also
support this headers object instead.