Use can-log for printing warnings
This uses can-log/dev/dev for logging the warning and removes the need
to check process.env.NODE_ENV (which may not be set), instead passing
down the isDevelopment flag that we already have.

This change removes the need for sinon, as can-test-helpers already
provides the log checking that we needed.