Provide the request object for H1 and H2
Previously I was only going to include the `request` property on the
ViewModel if it was an H1 connection. This was because H2 doesn't
actually have a "request", rather an object that are the headers.  I now
realize that omitting an expected property only in one scenario is
confusing. So this adds it back; in the case of H2 it will be the
headers object.