Prevent ssr memory leak
In SSR we need to replace the entire HTML element, otherwise the
AppViewModel will leak (as it gets set on the documentElement).

So this changes the strategy a little bit. Instead of moving the
head/body contents, we now go back to replacing the documentElement and
then moving head contents over.