Store nodeList on the Zone's data
This adds the nodeList to the Zone data rather than on the main export.
The reason is that by putting it on the main export multiple requests
will overwrite the previous version. Putting this on the Zone's data
object is the best place to put stuff specific to a request.